requirements.txt
streamlit==1.44.1
tensorflow==2.19.0
pillow==11.2.1
numpy==2.1.3
pandas==2.2.3
